# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RideTech API Documentation",
        "description": "Чистая интерактивная документация API для сервиса RideTech (вынесена отдельно от контроллеров)",
        "contact": {
            "email": "shavkat@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Основной сервер разработки"
        }
    ],
    "paths": {
...
        "/trips": {
            "get": {
                "tags": [
                    "Trips - Passenger"
                ],
                "summary": "Пассажир: Просмотр истории своих поездок",
                "description": "Возвращает пагинированный список всех поездок, созданных текущим авторизованным пассажиром.",
                "operationId": "getPassengerTripsHistory",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на странице (по умолчанию 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу поездки",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "in_progress",
                                "completed",
                                "cancelled"
                            ],
                            "example": "completed"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Фильтр по дате отправления (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-06-15"
                        }
                    },
                    {
                        "name": "driver_id",
                        "in": "query",
                        "description": "Фильтр по ID водителя",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос истории поездок",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "passenger_id": {
                                                        "description": "ID пассажира (users.id)",
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "driver_id": {
                                                        "description": "ID водителя (users.id)",
                                                        "type": "integer",
                                                        "example": 3,
                                                        "nullable": true
                                                    },
                                                    "car_id": {
                                                        "description": "ID автомобиля (cars.id)",
                                                        "type": "integer",
                                                        "example": 1,
                                                        "nullable": true
                                                    },
                                                    "origin": {
                                                        "type": "string",
                                                        "example": "Душанбе, ул. Рудаки 15"
                                                    },
                                                    "destination": {
                                                        "type": "string",
                                                        "example": "Худжанд, 19-й микрорайон"
                                                    },
                                                    "departure_time": {
                                                        "type": "string",
                                                        "example": "2026-06-15 14:00:00"
                                                    },
                                                    "preferences": {
                                                        "type": "string",
                                                        "example": "С кондиционером",
                                                        "nullable": true
                                                    },
                                                    "status": {
                                                        "description": "Текущий статус из TripStatus Enum",
                                                        "type": "string",
                                                        "example": "completed"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-06-12 12:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-06-12 12:45:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trips - Passenger"
                ],
                "summary": "Пассажир: Создание поездки",
                "description": "Позволяет авторизованному пассажиру создать новую заявку на поездку. По умолчанию выставляется статус PENDING.",
                "operationId": "createTrip",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "origin",
                                    "destination",
                                    "departure_time"
                                ],
                                "properties": {
                                    "origin": {
                                        "description": "Место отправления",
                                        "type": "string",
                                        "example": "Душанбе, ул. Рудаки 15"
                                    },
                                    "destination": {
                                        "description": "Место назначения",
                                        "type": "string",
                                        "example": "Худжанд, 19-й микрорайон"
                                    },
                                    "departure_time": {
                                        "description": "Время отправления в формате Y-m-d H:i:s",
                                        "type": "string",
                                        "example": "2026-06-15 14:00:00"
                                    },
                                    "preferences": {
                                        "description": "Предпочтения пассажира",
                                        "type": "string",
                                        "example": "С кондиционером, без курящих",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Поездка успешно создана, ожидайте водителя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Поездка успешно создана, ожидайте водителя"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "passenger_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "driver_id": {
                                                    "type": "integer",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "car_id": {
                                                    "type": "integer",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "origin": {
                                                    "type": "string",
                                                    "example": "Душанбе, ул. Рудаки 15"
                                                },
                                                "destination": {
                                                    "type": "string",
                                                    "example": "Худжанд, 19-й микрорайон"
                                                },
                                                "departure_time": {
                                                    "type": "string",
                                                    "example": "2026-06-15 14:00:00"
                                                },
                                                "preferences": {
                                                    "type": "string",
                                                    "example": "С кондиционером, без курящих",
                                                    "nullable": true
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-06-12 15:00:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-06-12 15:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован"
                    },
                    "422": {
                        "description": "Ошибка валидации входных данных"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/trips/{id}/cancel": {
            "post": {
                "tags": [
                    "Trips - Passenger"
                ],
                "summary": "Пассажир: Отмена поездки",
                "description": "Позволяет пассажиру отменить свою активную поездку. Статус изменяется на CANCELLED.",
                "operationId": "cancelTrip",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поездки (trips.id)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Поездка успешно отменена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Поездка успешно отменена"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "passenger_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "driver_id": {
                                                    "type": "integer",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "car_id": {
                                                    "type": "integer",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "origin": {
                                                    "type": "string",
                                                    "example": "Душанбе, ул. Рудаки 15"
                                                },
                                                "destination": {
                                                    "type": "string",
                                                    "example": "Худжанд, 19-й микрорайон"
                                                },
                                                "departure_time": {
                                                    "type": "string",
                                                    "example": "2026-06-15 14:00:00"
                                                },
                                                "preferences": {
                                                    "type": "string",
                                                    "example": "С кондиционером, без курящих",
                                                    "nullable": true
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "cancelled"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-06-12 15:00:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-06-12 15:10:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Невозможно отменить поездку (например, она уже в пути или завершена)"
                    },
                    "401": {
                        "description": "Неавторизован"
                    },
                    "404": {
                        "description": "Поездка не найдена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/trips/{id}/accept": {
            "post": {
                "tags": [
                    "Trips - Driver"
                ],
                "summary": "Водитель: Принять поездку",
                "description": "Привязывает водителя (driver_id) и его автомобиль (car_id) к поездке. Статус меняется на IN_PROGRESS.",
                "operationId": "acceptTrip",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поездки",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "car_id"
                                ],
                                "properties": {
                                    "car_id": {
                                        "description": "ID автомобиля водителя из таблицы cars",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Вы успешно приняли заказ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Вы успешно приняли заказ"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "passenger_id": {
                                                    "type": "integer",
                                                    "example": 8
                                                },
                                                "driver_id": {
                                                    "description": "ID текущего авторизованного водителя",
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "car_id": {
                                                    "description": "ID выбранного автомобиля из запроса",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "origin": {
                                                    "type": "string",
                                                    "example": "Канибадам, Центр"
                                                },
                                                "destination": {
                                                    "type": "string",
                                                    "example": "Худжанд, Шёлкокомбинат"
                                                },
                                                "departure_time": {
                                                    "type": "string",
                                                    "example": "2026-06-16 08:00:00"
                                                },
                                                "status": {
                                                    "description": "Статус меняется на in_progress",
                                                    "type": "string",
                                                    "example": "in_progress"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован"
                    },
                    "422": {
                        "description": "Ошибка валидации (car_id не существует в таблице cars)"
                    },
                    "404": {
                        "description": "Поездка не найдена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите ваш API токен в формате: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}